pyproject.toml
[project]
name = "big-pickle-blogging"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.103.2",
    "bcrypt>=3.2.0,<4.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-slugify>=8.0.4",
]
